restructuring: step 1
